requirements.txt
asgiref==3.3.1
Django==3.1.4
django-modeltranslation==0.16.1
djangorestframework==3.12.2
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
